combine components
