--- src/main/webapp/app/shared/metis/posting-reactions-bar/answer-post-reactions-bar/answer-post-reactions-bar.component.html
@@ -1,100 +1,0 @@
-<div class="d-flex flex-wrap gap-2 fs-xx-small align-items-center">
-    @for (reactionMetaData of reactionMetaDataMap | keyvalue; track reactionMetaData) {
-        @if (isEmojiCount) {
-            <div>
-                <button
-                    [disabled]="isReadOnlyMode"
-                    class="reaction-button clickable"
-                    [class.reaction-button--reacted]="reactionMetaData.value.hasReacted"
-                    (click)="updateReaction(reactionMetaData.key)"
-                    [ngbTooltip]="reactionMetaData.value.reactingUsers | reactingUsersOnPosting | async"
-                >
-                    <div class="emoji-container">
-                        <jhi-emoji [emoji]="reactionMetaData.key" />
-                        <span class="emoji-count" [class.reaction-count--reacted]="reactionMetaData.value.hasReacted">{{ reactionMetaData.value.count }}</span>
-                    </div>
-                </button>
-            </div>
-        }
-    }
-    <div [hidden]="isReadOnlyMode">
-        @if ((isAnyReactionCountAboveZero() && isEmojiCount) || !isEmojiCount) {
-            <ng-container>
-                <button
-                    class="reaction-button clickable px-2 fs-small"
-                    (click)="showReactionSelector = !showReactionSelector"
-                    cdkOverlayOrigin
-                    #trigger="cdkOverlayOrigin"
-                    [title]="'artemisApp.metis.chooseReaction' | artemisTranslate"
-                >
-                    <fa-icon [icon]="farSmile" [class.open-selector]="showReactionSelector" [title]="'artemisApp.metis.chooseReaction' | artemisTranslate" />
-                </button>
-                @if (!isReadOnlyMode) {
-                    <ng-template
-                        cdkConnectedOverlay
-                        [cdkConnectedOverlayHasBackdrop]="true"
-                        [cdkConnectedOverlayBackdropClass]="'cdk-overlay-transparent-backdrop'"
-                        [cdkConnectedOverlayOrigin]="trigger"
-                        [cdkConnectedOverlayOpen]="showReactionSelector"
-                        (backdropClick)="toggleEmojiSelect()"
-                    >
-                        <jhi-emoji-picker
-                            [emojisToShowFilter]="emojisToShowFilter"
-                            (emojiSelect)="selectReaction($event)"
-                            [categoriesIcons]="categoriesIcons"
-                            [recent]="selectedCourseEmojis"
-                        />
-                    </ng-template>
-                }
-            </ng-container>
-        }
-        @if (!isEmojiCount) {
-            @if (!isAnswerOfAnnouncement && (isAtLeastTutorInCourse || isAuthorOfOriginalPost)) {
-                <button
-                    class="reaction-button clickable px-2 fs-small resolve"
-                    id="toggleElement"
-                    [ngClass]="isAtLeastTutorInCourse || isAuthorOfOriginalPost ? 'clickable' : ''"
-                    (click)="toggleResolvesPost()"
-                >
-                    @if (posting.resolvesPost) {
-                        <fa-icon
-                            class="resolved"
-                            [icon]="faCheck"
-                            [ngbTooltip]="
-                                isAtLeastTutorInCourse || isAuthorOfOriginalPost
-                                    ? ('artemisApp.metis.unmarkAsResolvingAnswerTooltip' | artemisTranslate)
-                                    : ('artemisApp.metis.resolvingAnswerTooltip' | artemisTranslate)
-                            "
-                        />
-                    } @else {
-                        @if (isAtLeastTutorInCourse || isAuthorOfOriginalPost) {
-                            <fa-icon class="icon notResolved" [icon]="faCheck" [ngbTooltip]="'artemisApp.metis.markAsResolvingAnswerTooltip' | artemisTranslate" />
-                        }
-                    }
-                </button>
-            }
-            @if (mayEdit) {
-                <button class="reaction-button clickable px-2 fs-small edit" (click)="editPosting()" [ngbTooltip]="'artemisApp.metis.editPosting' | artemisTranslate">
-                    <fa-icon [icon]="faPencilAlt" cdkOverlayOrigin />
-                </button>
-            }
-            @if (mayDelete) {
-                <button class="reaction-button clickable px-2 fs-small delete">
-                    <jhi-confirm-icon
-                        iconSize="sm"
-                        (confirmEvent)="deletePosting()"
-                        [initialTooltip]="'artemisApp.metis.deleteAnswer' | artemisTranslate"
-                        [confirmTooltip]="'artemisApp.metis.confirmDeleteAnswer' | artemisTranslate"
-                    />
-                </button>
-            }
-            <button
-                class="reaction-button clickable px-2 fs-small bookmark"
-                (click)="bookmarkPosting()"
-                [ngbTooltip]="(posting.isSaved ? 'artemisApp.metis.post.removeBookmarkPost' : 'artemisApp.metis.post.bookmarkPost') | artemisTranslate"
-            >
-                <fa-icon [icon]="posting.isSaved ? faBookmark : farBookmark" cdkOverlayOrigin />
-            </button>
-        }
-    </div>
-</div>
