--- src/main/webapp/app/shared/metis/posting-reactions-bar/post-reactions-bar/post-reactions-bar.component.html
@@ -1,149 +1,0 @@
-<div class="d-flex flex-wrap gap-2 fs-xx-small align-items-center" style="width: max-content">
-    @if (hoverBar && sortedAnswerPosts.length === 0) {
-        <div>
-            <button class="reaction-button clickable reply-btn" (click)="isCommunicationPage ? openThread.emit() : openAnswerView()">
-                <fa-icon class="fa-xs align-self-center" [icon]="faArrowRight" />
-                <span jhiTranslate="artemisApp.conversationsLayout.threadSideBar.reply" class="emoji-count px-2"></span>
-            </button>
-        </div>
-    }
-    @if (!isCommunicationPage) {
-        @if (sortedAnswerPosts.length) {
-            <!-- collapse answers -->
-            @if (showAnswers) {
-                <div>
-                    <button class="reaction-button clickable collapse-answers-btn" [class.reaction-button--reacted]="showAnswers" (click)="closeAnswerView()">
-                        <jhi-emoji [emoji]="closeCrossId" />
-                        <span jhiTranslate="artemisApp.metis.collapseAnswers" class="emoji-count"></span>
-                    </button>
-                </div>
-            } @else {
-                <!-- expand answers -->
-                <div>
-                    <button class="reaction-button clickable expand-answers-btn" (click)="openAnswerView()">
-                        <fa-icon class="fa-xs align-self-center" [icon]="faArrowRight" />
-                        <span class="emoji-count">{{
-                            sortedAnswerPosts.length === 1
-                                ? ('artemisApp.metis.showSingleAnswer' | artemisTranslate)
-                                : ('artemisApp.metis.showMultipleAnswers' | artemisTranslate: { number: sortedAnswerPosts.length })
-                        }}</span>
-                    </button>
-                </div>
-            }
-        }
-    } @else {
-        @if (!isThreadSidebar) {
-            <!-- expand answers -->
-            @if (!showAnswers && sortedAnswerPosts.length) {
-                <div>
-                    <button class="reaction-button clickable expand-answers-btn" (click)="openThread.emit()">
-                        <fa-icon [icon]="faArrowRight" />
-                        <span class="emoji-count px-2">{{
-                            sortedAnswerPosts.length === 1
-                                ? ('artemisApp.metis.showSingleAnswer' | artemisTranslate)
-                                : ('artemisApp.metis.showMultipleAnswers' | artemisTranslate: { number: sortedAnswerPosts.length })
-                        }}</span>
-                    </button>
-                </div>
-            }
-        }
-    }
-    @for (reactionMetaData of reactionMetaDataMap | keyvalue; track reactionMetaData) {
-        @if (isEmojiCount) {
-            <div>
-                <button
-                    [disabled]="readOnlyMode"
-                    class="reaction-button clickable"
-                    [class.reaction-button--reacted]="reactionMetaData.value.hasReacted"
-                    (click)="updateReaction(reactionMetaData.key)"
-                    [ngbTooltip]="reactionMetaData.value.reactingUsers | reactingUsersOnPosting | async"
-                >
-                    <div class="emoji-container">
-                        <jhi-emoji [emoji]="reactionMetaData.key" />
-                        <span class="emoji-count" [class.reaction-count--reacted]="reactionMetaData.value.hasReacted">{{ reactionMetaData.value.count }}</span>
-                    </div>
-                </button>
-            </div>
-        }
-    }
-    <div [hidden]="readOnlyMode">
-        @if ((isAnyReactionCountAboveZero() && isEmojiCount) || !isEmojiCount) {
-            <ng-container>
-                <button
-                    class="reaction-button clickable px-2 fs-small"
-                    (click)="showReactionSelector = !showReactionSelector"
-                    cdkOverlayOrigin
-                    #trigger="cdkOverlayOrigin"
-                    [title]="'artemisApp.metis.chooseReaction' | artemisTranslate"
-                >
-                    <fa-icon [icon]="faSmile" [class.open-selector]="showReactionSelector" [title]="'artemisApp.metis.chooseReaction' | artemisTranslate" />
-                </button>
-
-                <ng-template
-                    cdkConnectedOverlay
-                    [cdkConnectedOverlayHasBackdrop]="true"
-                    [cdkConnectedOverlayBackdropClass]="'cdk-overlay-transparent-backdrop'"
-                    [cdkConnectedOverlayOrigin]="trigger"
-                    [cdkConnectedOverlayOpen]="showReactionSelector"
-                    (backdropClick)="toggleEmojiSelect()"
-                >
-                    @if (!readOnlyMode) {
-                        <jhi-emoji-picker
-                            [emojisToShowFilter]="emojisToShowFilter"
-                            (emojiSelect)="selectReaction($event)"
-                            [categoriesIcons]="categoriesIcons"
-                            [recent]="selectedCourseEmojis"
-                        />
-                    }
-                </ng-template>
-            </ng-container>
-        }
-
-        @if (!isEmojiCount && mayEdit) {
-            <button
-                class="reaction-button clickable px-2 fs-small edit"
-                [disabled]="readOnlyMode"
-                (click)="editPosting()"
-                [ngbTooltip]="'artemisApp.metis.editPosting' | artemisTranslate"
-            >
-                <fa-icon [icon]="faPencilAlt" cdkOverlayOrigin />
-            </button>
-        }
-        <jhi-post-create-edit-modal #createEditModal [posting]="posting" [isCommunicationPage]="isCommunicationPage" (isModalOpen)="isModalOpen.emit()" />
-        @if (!isEmojiCount && mayDelete) {
-            <button class="reaction-button clickable fs-small">
-                <jhi-confirm-icon
-                    iconSize="sm"
-                    (confirmEvent)="deletePosting()"
-                    [initialTooltip]="'artemisApp.metis.deleteAnswer' | artemisTranslate"
-                    [confirmTooltip]="'artemisApp.metis.confirmDeleteAnswer' | artemisTranslate"
-                />
-            </button>
-        }
-        @if (!isEmojiCount && (displayPriority === DisplayPriority.PINNED || canPin)) {
-            <button
-                class="reaction-button pin clickable fs-small"
-                [class.reaction-button--not-hoverable]="!canPin"
-                [class.reaction-button--reacted]="displayPriority === DisplayPriority.PINNED && canPin"
-                [disabled]="!canPin || readOnlyMode"
-                (click)="canPin && togglePin()"
-            >
-                <jhi-emoji class="fs-x-small" [emoji]="pinEmojiId" [ngbTooltip]="pinTooltip | artemisTranslate" [placement]="'auto'" />
-            </button>
-        }
-        @if (!isEmojiCount) {
-            <button
-                class="reaction-button clickable px-2 fs-small bookmark"
-                (click)="bookmarkPosting()"
-                [ngbTooltip]="(posting.isSaved ? 'artemisApp.metis.post.removeBookmarkPost' : 'artemisApp.metis.post.bookmarkPost') | artemisTranslate"
-            >
-                <fa-icon [icon]="posting.isSaved ? faBookmark : farBookmark" cdkOverlayOrigin />
-            </button>
-        }
-    </div>
-    <div>
-        @if (isEmojiCount && getShowNewMessageIcon()) {
-            <div jhiTranslate="global.generic.new" class="badge bg-secondary hideAfter5Seconds"></div>
-        }
-    </div>
-</div>
